Remove unique indexes and manually look for duplicate words

+++ app/controllers/lessons_controller.rb
@@ -69,17 +69,18 @@ class LessonsController < ApplicationController
   #   Content: { 'errors': ['Not Found'] }
   def update
     lesson = Lesson.find(params[:id])
+    errors = check_duplicates
+    unless errors.empty?
+      render json: { 'errors': errors }, status: :bad_request # 400
+      return
+    end
     Wordinfo.where(id: lesson.wordinfo_ids).destroy_all
     lesson.attributes = lesson_params
-    begin
-      unless lesson.save
-        render json: { 'errors': lesson.errors }, status: :bad_request # 400
-        return
-      end
-      render json: lesson.reload
-    rescue ActiveRecord::RecordNotUnique => e
-      render json: { 'errors': e.cause }, status: :bad_request # 400
+    unless lesson.save
+      render json: { 'errors': lesson.errors }, status: :bad_request # 400
+      return
     end
+    render json: lesson.reload
   end
 
   # DELETE /api/lesson/:id
@@ -128,5 +129,27 @@ class LessonsController < ApplicationController
         sentences_attributes: [:id, :context_sentence]
       ]
     )
+  end
+
+  def check_duplicates
+    errors = {}
+    lesson = params[:lesson]
+    wordinfos = lesson['wordinfos']
+    wordinfos_uniq = wordinfos.uniq! { |wordinfo| wordinfo['word'].downcase }
+    errors['wordinfos.word'] = ['has already been taken'] unless wordinfos_uniq.nil?
+    wordinfos.each do |wordinfo|
+      errors = check_duplicates_wordinfo_nested wordinfo, 'roots', 'word', errors
+      errors = check_duplicates_wordinfo_nested wordinfo, 'synonyms', 'word', errors
+      errors = check_duplicates_wordinfo_nested wordinfo, 'antonyms', 'word', errors
+      errors = check_duplicates_wordinfo_nested wordinfo, 'sentences', 'context_sentence', errors
+    end
+    errors
+  end
+
+  def check_duplicates_wordinfo_nested(wordinfo, model, attribute, errors)
+    if wordinfo[model] && !wordinfo[model].uniq! { |m| m[attribute].downcase }.nil?
+      errors["wordinfos.#{model}.#{attribute}"] = ['has already been taken']
+    end
+    errors
   end
 end
